Change the default launch options and directory structure
